backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
email-validator